backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.1.1
pymongo==4.8.0
bcrypt==4.1.2
PyJWT==2.8.0
